+++ index.php
@@ -126,11 +126,8 @@ if (isLoggedIn()) {
                 <div class="NaviLink"><a href="/?p=statistik" id="link_statistik">Serverstatistik</a></div>
                 <div class="NaviLink"><a href="/?p=regeln" id="link_regeln">Regeln</a></div>
                 <div class="NaviLink"><a href="/?p=hilfe" id="link_hilfe">Hilfe</a></div>
-                <div class="NaviLink"><a href="/?p=impressum" id="link_impressum">Impressum</a></div>
-                <div class="NaviLink"><a href="/?p=datenschutz" id="link_datenschutz">Datenschutzhinweise</a></div>
-                <div class="NaviLink"><a href="https://www.buymeacoffee.com/bratkartoffel" id="link_spenden"
-                                         target="_blank">Spenden</a>
-                </div>
+                <div class="NaviLink"><a href="/?p=impressum" id="link_impressum">Impressum / Datenschutz</a></div>
+                <div class="NaviSpacer"></div>
                 <div class="NaviLink"><a href="/actions/logout.php" id="link_logout">Abmelden</a></div>
             </div>
         </div>
@@ -183,11 +180,8 @@ if (isLoggedIn()) {
             <div class="NaviLink"><a href="/?p=anmelden" id="link_anmelden">Anmelden</a></div>
             <div class="NaviLink"><a href="/?p=registrieren" id="link_registrieren">Registrieren</a></div>
             <div class="NaviLink"><a href="/?p=regeln" id="link_regeln">Regeln</a></div>
-            <div class="NaviLink"><a href="/?p=impressum" id="link_impressum">Impressum</a></div>
-            <div class="NaviLink"><a href="/?p=datenschutz" id="link_datenschutz">Datenschutzhinweise</a></div>
-            <div class="NaviLink"><a href="https://www.buymeacoffee.com/bratkartoffel" id="link_spenden"
-                                     target="_blank">Spenden</a>
-            </div>
+            <div class="NaviSpacer"></div>
+            <div class="NaviLink"><a href="/?p=impressum" id="link_impressum">Impressum / Datenschutz</a></div>
         </div>
         <?php
     }